Update:update readme.md (#24)
* uodate readme

* uodate readme

* update gpu-num -> gpu-amount

* update readme

* update readme

* update readme

* update readme

* update readme
