+++ bitfusion_device_plugin/Makefile
@@ -5,7 +5,7 @@ IMAGE_REPO ?= docker.io/bitfusiondeviceplugin
 DEVICE_IMAGE_NAME ?= bitfusion-device-plugin
 WEBHOOK_IMAGE_NAME ?= bitfusion-webhook
 PKG_IMAGE_NAME ?= bitfusion-client
-IMAGE_TAG  ?= 0.2.1
+IMAGE_TAG  ?= 0.1.1
 K8S_PLATFORM ?= community
 
 export K8S_PLATFORM
