+++ bitfusion_device_plugin/example/pod-memory.yaml
@@ -15,8 +15,8 @@ spec:
       command: ["python /benchmark/scripts/tf_cnn_benchmarks/tf_cnn_benchmarks.py --local_parameter_device=gpu --batch_size=32 --model=inception3"]
       resources:
         limits:
-          bitfusion.io/gpu-num: 1
-          bitfusion.io/gpu-memory: 10000
+          bitfusion.io/gpu-amount: 1
+          bitfusion.io/gpu-memory: 8000M
       volumeMounts:
         - name: code
           mountPath: /benchmark
