+++ bitfusion_device_plugin/webhook/deployment/bitfusion-injector.yaml
@@ -21,7 +21,7 @@ spec:
           imagePullPolicy: IfNotPresent
           env:
             - name: TOTAL_GPU_MEMORY
-              value: 16000
+              value: "0"
           args:
           - -sidecarCfgFile=/etc/webhook/config/sidecarconfig.yaml
           - -bitfusionClientConfig=/etc/webhook/bitfusion-client-config/bitfusion-client-config.yaml
