+++ bitfusion_device_plugin/webhook/pkg/webhook/util.go
@@ -108,8 +108,21 @@ func updateContainer(targets, source []corev1.Container, basePath string, bfClie
 		})
 
 		//container.Env = append(container.Env, source[0].Env...)
-		env := corev1.EnvVar{Name: "LD_LIBRARY_PATH", Value: bfClientConfig.EnvVariable}
-		container.Env = append(container.Env, env)
+		index := -1
+		for i := range container.Env {
+			glog.Infof("container.Env[i].Name = %s", container.Env[i].Name)
+			if container.Env[i].Name == "LD_LIBRARY_PATH" {
+				index = i
+				glog.Infof("index = %d", index)
+			}
+		}
+		if index != -1 {
+			env := corev1.EnvVar{Name: "LD_LIBRARY_PATH", Value: bfClientConfig.EnvVariable + ":" + container.Env[index].Value}
+			container.Env[index] = env
+		} else {
+			env := corev1.EnvVar{Name: "LD_LIBRARY_PATH", Value: bfClientConfig.EnvVariable}
+			container.Env = append(container.Env, env)
+		}
 		//env = corev1.EnvVar{Name: "PATH", Value: bfClientConfig.BinaryPath + ":$PATH"}
 		//container.Env = append(container.Env, env)
 		patches = append(patches, patchOperation{
@@ -265,8 +278,8 @@ func updateInitContainersResources(target, added []corev1.Container) []corev1.Co
 	return added
 }
 
-// CopySecret copies a secret to target namespace
-func CopySecret(namespace *string) error {
+// copySecret copies a secret to target namespace
+func copySecret(namespace *string) error {
 	name := "bitfusion-secret"
 	config, err := rest.InClusterConfig()
 	if err != nil {
@@ -454,13 +467,14 @@ func mutationRequired(ignoredList []string, metadata *metav1.ObjectMeta) bool {
 
 	// Determine whether to perform mutation based on annotation for the target resource
 	var required bool
+	injectionStatus = ""
 	if strings.ToLower(status) == "injected" {
 		required = false
 	} else {
 		switch strings.ToLower(annotations[admissionWebhookAnnotationInjectKey]) {
 		default:
 			required = false
-		case "y", "yes", "true", "on":
+		case "y", "yes", "true", "on", "all":
 			required = true
 		case bitFusionOnlyInjection:
 			injectionStatus = bitFusionOnlyInjection
